refactor MVC #89
